+++ api/src/test/java/org/grnet/cat/api/registry/RegistryCodelistEndpointTest.java
@@ -6,6 +6,7 @@ import io.restassured.http.ContentType;
 import org.grnet.cat.api.KeycloakTest;
 import org.grnet.cat.api.endpoints.registry.RegistryCodelistEndpoint;
 import org.grnet.cat.dtos.InformativeResponse;
+import org.grnet.cat.dtos.registry.RelationResponse;
 import org.grnet.cat.dtos.registry.codelist.*;
 import org.junit.jupiter.api.Test;
 
@@ -178,6 +179,45 @@ public class RegistryCodelistEndpointTest extends KeycloakTest {
                 .oauth2(getAccessToken("alice"))
                 .contentType(ContentType.JSON)
                 .get("/motivation-types/{id}", "pid_graph:5AF642D8")
+                .then()
+                .assertThat()
+                .statusCode(403)
+                .extract()
+                .as(InformativeResponse.class);
+
+        assertEquals("You do not have permission to access this resource.", error.message);
+    }
+
+
+    @Test
+    public void getRelation() {
+
+        register("admin");
+
+        var response = given()
+                .auth()
+                .oauth2(getAccessToken("admin"))
+                .contentType(ContentType.JSON)
+                .get("/relations/{id}", "dcterms:isRequiredBy")
+                .then()
+                .assertThat()
+                .statusCode(200)
+                .extract()
+                .as(RelationResponse.class);
+
+        assertEquals(response.id, "dcterms:isRequiredBy");
+    }
+
+    @Test
+    public void getRelationNotPermitted() {
+
+        register("alice");
+
+        var error = given()
+                .auth()
+                .oauth2(getAccessToken("alice"))
+                .contentType(ContentType.JSON)
+                .get("/relations/{id}", "dcterms:isRequiredBy")
                 .then()
                 .assertThat()
                 .statusCode(403)